NXP-4399: VCS: more efficient update for multi-valued or complex properties

+++ nuxeo-core-storage-sql/nuxeo-core-storage-sql/src/main/java/org/nuxeo/ecm/core/storage/sql/Mapper.java
@@ -1477,8 +1477,12 @@ public class Mapper {
      */
     public void updateCollectionRows(CollectionFragment fragment)
             throws StorageException {
+        if (!fragment.isDirty())  {
+            return;
+        }
         deleteFragment(fragment);
         insertCollectionRows(fragment);
+        fragment.setDirty(false);
     }
 
     /**
